requirements.txt
argparse>=1.4.0
subprocess>=3.2.7
PyYAML>=3.11
regex>=2018.01.10
HTSeq>=0.6.1p1
pyfaidx>=0.2.7
statsmodels>=0.6.1
pysam>=0.9.1.4
svgwrite>=1.1.6
numpy>=1.11.1



